Add files via upload
Solution for:
There are two arrays of integers arr1 and arr2. One move is defined as an increment or decrement by one of each digit of the items in the array. How many moves will it take to match the array arr1 with the array arr2? No reordering of the digits is allowed.